+++ docs/fr_FR/Changelog.rst
@@ -1,6 +1,9 @@
 ChangeLog
 =========
 
+- Interne: Parser: 'Batterie-Pourcent' => '0001-01-0021'.
+- Interne: install/DB cmd: 'Batterie-Pourcent' => '0001-01-0021'
+
 221105-BETA-1
 -------------